+++ main/src/main/java/net/oneandone/stool/configuration/StoolConfiguration.java
@@ -116,6 +116,9 @@ public class StoolConfiguration {
     /** Path to docker unix domain socket */
     public String docker;
 
+    /** Path available to system application containers */
+    public String systemExtras;
+
     public StoolConfiguration() {
         portFirst = 9000;
         portLast = 9999;
@@ -139,6 +142,7 @@ public class StoolConfiguration {
         search = "";
         quota = 0;
         docker = "/var/run/docker.sock";
+        systemExtras = "";
     }
 
     public static FileNode configurationFile(FileNode home) {